load commit

+++ 0x00-hello_world/6-size.c
@@ -10,13 +10,13 @@ int main(void)
 	int b;
 	long int c;
 	long long d;
-	int e;
+	float e;
 
-	printf("Size of a char: %d byte(s)\n", sizeof(a));
-	printf("Size of an int: %d byte(s)\n", sizeof(b));
-	printf("Size of a long int: %d byte(s)\n", sizeof(c));
-	printf("Size of a long long: %d byte(s)\n", sizeof(d));
-	printf("Size of a float: %d byte(s)\n", sizeof(e));
+	printf("Size of a char: %ld byte(s)\n", sizeof(a));
+	printf("Size of an int: %ld byte(s)\n", sizeof(b));
+	printf("Size of a long int: %ld byte(s)\n", sizeof(c));
+	printf("Size of a long long: %ld byte(s)\n", sizeof(d));
+	printf("Size of a float: %ld byte(s)\n", sizeof(e));
 
 	return (0);
 }
